# [SPARK-22032][PySpark] Speed up StructType conversion