--- a/python/pyspark/sql/types.py
+++ b/python/pyspark/sql/types.py
@@ -621,7 +621,7 @@ class StructType(DataType):
             # it's already converted by pickler
             return obj
         if self._needSerializeAnyField:
-            values = [f.fromInternal(v) if n else v 
+            values = [f.fromInternal(v) if n else v
                       for f, v, n in zip(self.fields, obj, self._needConversion)]
         else:
             values = obj