@HyukjinKwon 2017-09-16
How many functions call does it save and much it improves? I'd not bother fixing this.