@HyukjinKwon 2017-09-16
To be honest, it looks too trivial that I won't bother.