@HyukjinKwon 2017-09-16
Did it save 6~7% of the total execution time?