@HyukjinKwon 2017-09-16
Okay, then let's go ahead then. Let'd add some numbers in the PR description.