@maver1ck 2017-09-16
I added benchmark for this code