@ueshin 2017-09-18
A late LGTM. Btw, can we use the same idea for `MapType`?