@maver1ck 2017-09-18
@ueshin 
I think that for Maptype this is not a solution because every key / value of MapType is the same type so we need conversion for all entries or for nothing