@HyukjinKwon 2017-09-18
We can split two `needConversion` for key and value only and save key conversion or value conversion call though?